create webView component
