+++ app/routes/reactWall/reactWallScene.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
 import Header from '../../components/header.js';
+import WebView from '../../components/webViewComponent.js';
 import PropTypes from 'prop-types';
 
 class ReactWallSceen extends Component {
@@ -12,11 +13,7 @@ class ReactWallSceen extends Component {
             <View style={{ flex: 1 }}>
                 <Header leftIcon="bars" navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')} title="React Wall" />
                 <View style={styles.container}>
-                    <TouchableOpacity onPress={() => this.props.navigation.navigate('WebView', { setUrl: 'https://www.google.pl' })} >
-                        <Text style={styles.welcome}>
-                          Welcome to ReactWall!
-                        </Text>
-                    </TouchableOpacity>
+                    <WebView setUrl={'http://google.pl'} onHandlePress={() => this.props.navigation.navigate('WebView', { setUrl: 'https://www.google.pl' })}/>
                 </View>
             </View>
         );
