+++ app/routes/webView/webViewScene.js
@@ -1,20 +1,12 @@
 import React, { Component, PropTypes } from 'react';
 import { View, StyleSheet, Dimensions, WebView, Linking } from 'react-native';
 import Header from '../../components/header.js';
-import { connect } from 'react-redux';
 const window = Dimensions.get('window');
 
 class webViewScene extends Component {
     static propTypes = {
-        navigation: PropTypes.object,
-        webViewEnable: PropTypes.bool
+        navigation: PropTypes.object
     };
-    componentDidMount () {
-        const { params } = this.props.navigation.state;
-        if (this.props.webViewEnable) {
-            Linking.openURL(params.setUrl);
-        }
-    }
     render () {
         const { params } = this.props.navigation.state;
         const url = params.setUrl;
@@ -32,20 +24,16 @@ class webViewScene extends Component {
         if (arrLen > 2) {
             domain = splitArr[arrLen - 2] + '.' + splitArr[arrLen - 1];
         }
-        if (this.props.webViewEnable) {
-            return (Linking.openURL(url));
-        } else {
-            return (
-                <View style={styles.webViewStyle}>
-                    <Header leftIcon="arrow-left" navigatorLeft={() => this.props.navigation.goBack()} title={domain}
-                      renderRight rightIcon="external-link" navigatorRight={() => Linking.openURL(url)} />
-                    <WebView
-                      style={styles.webViewStyle}
-                      source={{ url }}
-                    />
-                </View>
-            );
-        }
+        return (
+            <View style={styles.webViewStyle}>
+                <Header leftIcon="arrow-left" navigatorLeft={() => this.props.navigation.goBack()} title={domain}
+                  renderRight rightIcon="external-link" navigatorRight={() => Linking.openURL(url)} />
+                <WebView
+                  style={styles.webViewStyle}
+                  source={{ url }}
+                />
+            </View>
+        );
     }
 }
 
@@ -59,10 +47,4 @@ const styles = StyleSheet.create({
     }
 });
 
-const mapStateToProps = ({ switcher }) => {
-    return {
-        webViewEnable: switcher.webViewEnable
-    };
-};
-
-export default connect(mapStateToProps)(webViewScene);
+export default webViewScene;
